third_party/WebKit/LayoutTests/http/tests/origin_trials/webexposed/installedapp-origin-trial-interfaces-script-added-expected.txt

This is a testharness.js-based test.
FAIL InstalledApp related properties is not on interfaces before adding trial token via script. assert_equals: Property getInstalledRelatedApps exists on Navigator expected false but got true
PASS InstalledApp related properties is on interfaces after adding trial token via script.
Harness: the test ran to completion.
